Fix jest test failures. This is caused by unrecognized ES6 import statements in OpenLayers that jest was pulling in. As this is already a known issue and our code under test doesn't touch OpenLayers, and in the interest of "sweeping this under the carpet" once again, the new OL style helpers have been moved to their own dedicated module.
