+++ jest.config.js
@@ -3,6 +3,9 @@ module.exports = {
     transform: {
         ".(ts|tsx)": "ts-jest"
     },
+    transformIgnorePatterns: [
+        "node_modules/ol"
+    ],
     testRegex: "(/test/.*|\\.(test|spec))\\.(ts|tsx|js)$",
     moduleFileExtensions: [
         "ts",
