+++ src/components/layer-manager/manage-layers.tsx
@@ -1,10 +1,11 @@
 import * as React from "react";
 import { tr } from "../../api/i18n";
-import { ILayerInfo, GenericEvent, IVectorFeatureStyle } from "../../api/common";
-import { ITreeNode, Tree, Button, Intent, ButtonGroup, Card, Icon, Switch, NonIdealState, Slider, Collapse } from '@blueprintjs/core';
+import { ILayerInfo } from "../../api/common";
+import { Button, Intent, ButtonGroup, Card, Icon, Switch, NonIdealState, Slider, Collapse } from '@blueprintjs/core';
 import { BlueprintSvgIconNames } from 'src/constants';
 import { strIsNullOrEmpty } from "../../utils/string";
 import { VectorStyleEditor } from '../vector-style-editor';
+import { IVectorFeatureStyle } from '../../api/ol-style-helpers';
 
 interface IManageLayerItemProps {
     layer: ILayerInfo;
@@ -71,14 +72,14 @@ const ManageLayerItem = (props: IManageLayerItemProps) => {
     }
     const isWmsLegendOpen = !strIsNullOrEmpty(wmsLegendUrl);
     return <Card key={layer.name}>
-        <Switch checked={layer.visible} onChange={e => onSetVisibility(layer.name, !layer.visible)} labelElement={<><Icon icon={iconName} /> {layer.name}</>} />
+        <Switch checked={layer.visible} onChange={() => onSetVisibility(layer.name, !layer.visible)} labelElement={<><Icon icon={iconName} /> {layer.name}</>} />
         <p>Opacity</p>
         <Slider min={0} max={1.0} stepSize={0.01} value={layer.opacity} onChange={e => onSetOpacity(layer.name, e)} />
         <ButtonGroup>
-            <Button title={tr("LAYER_MANAGER_TT_MOVE_UP", locale)} intent={Intent.PRIMARY} icon="caret-up" onClick={(e: any) => onMoveLayerUp(layer.name)} disabled={!canMoveUp} />
-            <Button title={tr("LAYER_MANAGER_TT_MOVE_DOWN", locale)} intent={Intent.PRIMARY} icon="caret-down" onClick={(e: any) => onMoveLayerDown(layer.name)} disabled={!canMoveDown} />
-            <Button title={tr("LAYER_MANAGER_TT_ZOOM_EXTENTS", locale)} intent={Intent.SUCCESS} icon="zoom-to-fit" onClick={(e: any) => onZoomToBounds(layer.name)} disabled={!canZoom} />
-            <Button title={tr("LAYER_MANAGER_TT_REMOVE", locale)} intent={Intent.DANGER} icon="trash" onClick={(e: any) => onRemoveLayer(layer.name)} />
+            <Button title={tr("LAYER_MANAGER_TT_MOVE_UP", locale)} intent={Intent.PRIMARY} icon="caret-up" onClick={() => onMoveLayerUp(layer.name)} disabled={!canMoveUp} />
+            <Button title={tr("LAYER_MANAGER_TT_MOVE_DOWN", locale)} intent={Intent.PRIMARY} icon="caret-down" onClick={() => onMoveLayerDown(layer.name)} disabled={!canMoveDown} />
+            <Button title={tr("LAYER_MANAGER_TT_ZOOM_EXTENTS", locale)} intent={Intent.SUCCESS} icon="zoom-to-fit" onClick={() => onZoomToBounds(layer.name)} disabled={!canZoom} />
+            <Button title={tr("LAYER_MANAGER_TT_REMOVE", locale)} intent={Intent.DANGER} icon="trash" onClick={() => onRemoveLayer(layer.name)} />
             {extraActions}
         </ButtonGroup>
         {isWms && <Collapse isOpen={isWmsLegendOpen}>
@@ -129,14 +130,6 @@ export const ManageLayers = (props: IManageLayersProps) => {
     } = props;
     const [layers, setLayers] = React.useState(props.layers);
     const [wmsLegendUrl, setWmsLegendUrl] = React.useState<string | undefined>(undefined);
-    const onToggleWmsLegend = (action: (res?: number) => string) => {
-        if (wmsLegendUrl) {
-            setWmsLegendUrl(undefined);
-        } else {
-            const url = action(currentResolution);
-            setWmsLegendUrl(url);
-        }
-    };
     React.useEffect(() => {
         setLayers(props.layers);
     }, [props.layers]);
